+++ bimap/bimap.go
