+++ fifo/queue.go
